Add a BMX160 init helper function to make this library work out-of-the-box
Until now, the BMX160 library was not working out-of-the-box. As said in
the README.md, it required a manual change. This adds an unnecessary step
that remains hidden until several days of debugging are spent by adopters.

The solution is a new helper function based on bmi160_init but initializes
BMX160 with its proper CHIPID(0xD8).

Signed-off-by: Dimitar Tomov <dimi@edgeimpulse.com>